fix(analytics):fix analytics code climate issues
 - Create analytics_helpers file to contain common analytics methods

 - Format analytics methods to use analytics_helper methods

[Delivers #161296049]
